LPS-53683 moves java
