+++ modules/apps/search/search-web/src/META-INF/resources/open_search.jspf
@@ -28,7 +28,7 @@ else {
 	primarySearch = portalPreferences.getValue(PortletKeys.SEARCH, "primary-search", StringPool.BLANK);
 }
 
-List<OpenSearch> openSearchInstances = com.liferay.portlet.search.util.SearchUtil.getOpenSearchInstances(primarySearch);
+List<OpenSearch> openSearchInstances = com.liferay.search.web.util.SearchUtil.getOpenSearchInstances(primarySearch);
 
 List<String> modelResources = new ArrayList<String>();
 
@@ -70,7 +70,7 @@ for (int i = 0; i < openSearchInstances.size(); i++) {
 
 	int inactiveGroupsCount = GroupLocalServiceUtil.searchCount(themeDisplay.getCompanyId(), null, null, groupParams);
 
-	Tuple tuple = com.liferay.portlet.search.util.SearchUtil.getElements(xml, openSearch.getClassName(), inactiveGroupsCount);
+	Tuple tuple = com.liferay.search.web.util.SearchUtil.getElements(xml, openSearch.getClassName(), inactiveGroupsCount);
 
 	List<Element> resultRows = (List<Element>)tuple.getObject(0);
 	int totalRows = (Integer)tuple.getObject(1);
@@ -128,9 +128,10 @@ for (int i = 0; i < openSearchInstances.size(); i++) {
 					String summary = element.elementText("summary");
 					String entryClassName = element.elementText("entryClassName");
 					long entryClassPK = GetterUtil.getLong(element.elementText("entryClassPK"));
+					String[] queryTerms = StringUtil.split(element.elementText("queryTerms"), StringPool.COMMA_AND_SPACE);
 					%>
 
-					<a class="entry-title" href="<%= com.liferay.portlet.search.util.SearchUtil.getSearchResultViewURL(renderRequest, renderResponse, entryClassName, entryClassPK, searchDisplayContext.isViewInContext(), currentURL) %>">
+					<a class="entry-title" href="<%= com.liferay.search.web.util.SearchUtil.getSearchResultViewURL(renderRequest, renderResponse, entryClassName, entryClassPK, searchDisplayContext.isViewInContext(), currentURL) %>">
 						<%= SearchUtil.highlight(HtmlUtil.escape(entryTitle), queryTerms) %>
 					</a>
