+++ modules/apps/search/search-web/src/com/liferay/search/web/portlet/dependencies/facet_configuration.json
